refactored to use mappers and other onion structure

+++ src/Dtos/ReportDtos/RecentQuizAttemptDto.cs
@@ -1,15 +1,10 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
 namespace BrainThrust.src.Dtos.ReportDtos
 {
     public class RecentQuizAttemptDto
     {
         public int UserId { get; set; }
         public int QuizId { get; set; }
-        public int TotalScore { get; set; }
+        public double TotalScore { get; set; }
         public bool IsPassed { get; set; }
         public DateTime Created { get; set; }
     }
