+++ src/Dtos/ReportDtos/TopScorerDto.cs
@@ -1,13 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
 namespace BrainThrust.src.Dtos.ReportDtos
 {
     public class TopScorerDto
     {
         public int UserId { get; set; }
-        public int TotalScore { get; set; }
+        public double TotalScore { get; set; }
     }
 }
